+++ TransferObject/ChamCong/LoaiCa.cs
@@ -10,13 +10,13 @@ namespace TransferObject
     {
         public int IdLoaiCa { get; set; }
         public string TenLoaiCa { get; set; }
-        public float HeSo { get; set; }
+        public double HeSo { get; set; }
         public TimeSpan GioBatDau { get; set; } 
         public TimeSpan GioKetThuc { get; set; } 
 
         public LoaiCa() { }
 
-        public LoaiCa(int id, string tenLoaiCa, TimeSpan gioBatDau, TimeSpan gioKetThuc, float heSo)
+        public LoaiCa(int id, string tenLoaiCa, TimeSpan gioBatDau, TimeSpan gioKetThuc, double heSo)
         {
             this.IdLoaiCa = id;
             this.TenLoaiCa = tenLoaiCa;
